Refactor database
